requirements.txt
aliyun-python-sdk-core==2.13.15
aliyun-python-sdk-sts==3.0.1
APScheduler==3.6.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
cssselect==1.1.0
elasticsearch==7.10.1
Django==3.0.5
djangorestframework==3.11.1
gevent==20.4.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
javaobj-py3==0.4.1
jmespath==0.9.5
JPype1==0.7.4
kafka-python==2.0.1
lxml==4.5.2
numpy==1.19.2
oss2==2.11.0
pandas==1.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
pycryptodomex==3.9.8
pyjks==20.0.0
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
schedule==0.6.0
six==1.15.0
twofish==0.3.0
tzlocal==2.1
urllib3==1.25.9
yagmail==0.14.245
flask==2.0.1
mongoengine==0.23.1
flask-caching==1.10.1
click ==8.0.1
SQLAlchemy==1.3.22